First macro review
This first commit is an initial major cleaning of the roles, removing
what is not used and fixing bits and pieces to get a working solution on
a three machines setup after the playbook run.
This is just the first step of many to come, we're far from a production
ready solution.