Reference k8s token directly from master node
This commit makes the worker nodes that need to join the master k8s node
to rely on the token stored as a fact from the master.
Earlier every time we needed it, we executed the extraction on each
worker, executing the command as many times as the number of worker
nodes.
Now the token is extracted just one time and referenced from the workers
as an hostvar.