Fix graylog playbook var and files
This huge commit removes unused stuff and moves the variables in the
correct location, so under the group_vars inventory directory.
It makes also a general review to some common tasks, like the hosts file
creation and the tuned activation.
It finally fixes all the anomalies listed by the ansible-linter.