Add support for users in k8s
This commit adds the support for users creation inside the Kubernetes
role. It relies on the k8s_users list, for each user this should be
declared:

k8s_users:
  - name: pod-viewer
    namespace: default
    role_name: pod-viewer-role
    role_rules_apigroups: '""'
    role_rules_resources: '"pods","pods/exec","pods/log"'
    role_rules_verbs: '"*"'
    rolebinding_name: pod-viewer-rolebinding
    cert_path: 'kubernetes/pod-viewer/ssl'
    cert_expire_days: 3650
    update_kube_config: true

It will create a namespace (if it does not exists), a role and a role
binding with specific permissions enabled for the user.
It will alse create the local ssl keys under cert_path and optionally
enable the new configuration inside the .kube/config file.