New tasks in main.yml (one for all .tf file) and new vars
This commit features a new .j2 template for each "family" of
hosts that will be deployed. All the templates are essentially
the same and will be in the future updated if necessary.
Also, main.yml has been modified to not use absolute path as
for destination of final .tf files.
The new vars added are the following (and must be added to the
inventory):
- vm_size: "Azure tier for the vm (standard_b,_e,_f, etc.);
- osdisk_size_ size for the os disk (minimum required by Azure
  is 30GB both for CentOS and Ubuntu images);
- disk[1]_size: size for the additional data disk. Right now
  j2 templates need to be edited if more than 1 data disk is
  needed (see apache.tf.j2);