Use linux_virtual_machine for terraform VMS
This commit replaces the old and full of complications terraform module
azurerm_virtual_machine with the more compatible and usable
azurerm_linux_virtual_machine. In this way disk encryption is smoother
and also user privileges are set up more coherently and in a readable
form for the final user that looks at the terraform .tf files.