Support groups exclusions for terraform vms
To generate the vms_<groupname>.tf files we start by looking to all the
groups except 'all' and 'ungrouped'. This is good if you have each
machine part of just one group, but when other groups are declared
for various reasons and a machine appears in more than one, until this
commit it was impossible to have the machine declared just one time.
This commit introduces the variable named azure_vms_excluded_groups that
will containe all the groups that will be excluded when generating the
vms_<groupname>.tf files.