Change group logic for vms terraform files
This commit changes the logic behing how the vms_*.tf files are
generated.
Previously we were considering all the groups, excluding them using the
azure_vms_excluded_groups list variable. Since the list can become huge
and the complexity of the groups declaration can as well become a
problem, we inverted the logic, using a variable named azure_vms_groups
in which the user will need to specify the list of the groups from which
the various vms_*.tf files will be generated.
By default the value of azure_vms_groups is 'all'.