Graylog: separate hosts groups by scope
This commit reviews the entire Graylog playbook (and role) so that it
will be to totally separate the hosts by their scope.
For each scope comes a specific group:

- graylog_mongodb: hosts deploying MongoDB cluster;
- graylog_elastic: hosts deploying Elasticsearch cluster;
- graylog_server: hosts deploying Graylog server;
- graylog_frontend: hosts deployinh Graylog VIP (NGinx/Keepalived);

All the hosts will be grouped under a specific macro group declared as
follows:

[graylog:children]
graylog_mongodb
graylog_elastic
graylog_server
graylog_frontend

The main playbook will configure the various roles on the different
groups so that it is now possible to split the entire Graylog setup
between single purpose machine groups.