Terraform - Fix VM extensions with multiple disks
Updated how the Terraform VM extension code will be generated when multiple
disks are defined in the inventory for a single VM. Instead of generating one
extension for each disk (which appears to be unsupported from Azure), now a
single extension will be generated which provide partitioning, formatting and
mounting of every disks specified in the inventory.
Also, a sort of protection was added in order to avoid disk disruption if the
extension will run on an already deployed VM: now disk initialization will be
skipped if the disk appears to be already initialized.
Added a method to specify a custom name for the extension.